Make mypy happy

+++ backend/xpark/logic/reservations.py
@@ -109,7 +109,7 @@ def calculate_booking_price(
     price = res["price"]
     delta = end_time - start_time
     hours = delta.days * 24 + delta.seconds / 3600
-    return hours * price
+    return float(hours * price)
 
 
 def create_reservation(
